requirements.txt
python-telegram-bot==20.3
asyncpg==0.27.0
apscheduler==3.10.1
python-dotenv==1.0.0
python-dateutil==2.8.2